About Page fix
fixes to about page to make sure it works properly with navbar and everything... also moved around login page and created its separate folder for me to work on
